Fix shellcheck violations; remove trailing semis
Most of the issues flagged were about missing quotes, but there were
also some cases of unneeded uses of `$?`, which is easy to break on
refactoring. It's much better to merge the `if` with the command so
they are not broken apart by some other command.

Trailing semicolons are also not a shellscript idiom, so that was
removed.